fix arc ut test
